Fix test script: generate FP8 tensors via BF16 cast

+++ scripts/test_triton_v3.py
@@ -29,13 +29,13 @@ print(f"Testing triton_fused_v3 with T={T}, E={E}, H={H}, I={I}")
 out = run(
     routing_logits=torch.randn(T, 256, device=device, dtype=torch.float32),
     routing_bias=torch.zeros(256, device=device, dtype=torch.float32),
-    hidden_states=torch.randn(T, H, device=device, dtype=torch.float8_e4m3fn),
+    hidden_states=torch.randn(T, H, device=device, dtype=torch.bfloat16).to(torch.float8_e4m3fn),
     hidden_states_scale=torch.ones(H // BLOCK, T, device=device, dtype=torch.float32),
-    gemm1_weights=torch.randn(E, 2 * I, H, device=device, dtype=torch.float8_e4m3fn),
+    gemm1_weights=torch.randn(E, 2 * I, H, device=device, dtype=torch.bfloat16).to(torch.float8_e4m3fn),
     gemm1_weights_scale=torch.ones(
         E, 2 * I // BLOCK, H // BLOCK, device=device, dtype=torch.float32
     ),
-    gemm2_weights=torch.randn(E, H, I, device=device, dtype=torch.float8_e4m3fn),
+    gemm2_weights=torch.randn(E, H, I, device=device, dtype=torch.bfloat16).to(torch.float8_e4m3fn),
     gemm2_weights_scale=torch.ones(
         E, H // BLOCK, I // BLOCK, device=device, dtype=torch.float32
     ),
